pyproject.toml
[project]
name = "fake-news-detector"
version = "0.1.0"
description = "Fake News Detection"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "fastapi>=0.121.1",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "lingua-language-detector>=2.1.1",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "pydantic-settings>=2.12.0",
    "pymongo>=4.15.4",
    "pytest>=9.0.1",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "shap>=0.49.1",
    "torch>=2.9.1",
    "transformers>=4.57.3",
    "uvicorn>=0.38.0",
    "wordcloud>=1.9.4",
]
